requirements.txt
julep
fastapi 
uvicorn
pyyaml